requirements.txt
asgiref==3.8.1
boto3==1.37.28
botocore==1.37.28
CacheControl==0.14.2
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
idna==3.10
iso3166==2.1.1
jmespath==1.0.1
mapbox==0.18.1
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
polyline==2.0.2
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
